requirements.txt
streamlit==1.28.1
langchain==0.0.350
openai==1.3.8
pypdf==3.17.1
python-dotenv==1.0.0
tiktoken==0.5.1


